Add register extension. Remove garbage.
