+++ simphony/engine/__init__.py
@@ -4,11 +4,57 @@ This module is dynamicaly populated at import with the
 registered plugins modules. Plugins modules need to be
 registered at the 'simphony.engine' entry point.
 """
-from ..extension import get_engine_manager
-
+from .extension import ABCEngineExtension
+from .extension import EngineInterface
+from .extension import EngineManager
 
 __all__ = ['get_supported_engines',
-           'get_supported_engine_names']
+           'get_supported_engine_names',
+           'ABCEngineExtension', 'EngineInterface']
+
+
+# TODO: Use an application server and put this in app context.
+# Wrapper manager class.
+_ENGINE_MANAGER = EngineManager()
+
+
+def get_engine_manager():
+    """Get the engine manager instance."""
+    return _ENGINE_MANAGER
+
+
+def set_engine_manager(manager):
+    """Use an arbitrary engine manager for testing.
+
+    Parameters
+    ----------
+    manager: EngineManager
+        an instance of engine manager
+    """
+    global _ENGINE_MANAGER
+    _ENGINE_MANAGER = manager
+
+
+def create_wrapper(cuds, engine_name, engine_interface=None):
+    """Create a wrapper to the given engine.
+
+    Parameters
+    ----------
+    cuds: CUDS
+        a cuds object which contains model information
+    engine_name: str
+        name of the underlying engine to launch the simulation with
+    engine_interface: engine.EngineInterface
+        the interface to the engine, internal or fileio
+
+    Returns
+    -------
+    wrapper: engine.ABCEngineExtension
+        an engine wrapper instance
+    """
+    return get_engine_manager().create_wrapper(cuds,
+                                               engine_name,
+                                               engine_interface)
 
 
 def get_supported_engine_names():
@@ -42,11 +88,8 @@ def load_engine_extentions():
         namespace='simphony.engine',
         invoke_on_load=False)
     extensions = {}
-    engine_manager = get_engine_manager()
     for ext in mgr.extensions:
         extensions[ext.name] = ext.plugin
-        # Load engine metadata
-        engine_manager.load_metadata(ext.plugin)
     return extensions
